Rename problem files and order directory readme
